test(cli): split watch tests
